requirements.txt
Django>=3.0,<3.1
python-decouple # ==3.3
celery>=4.3,<4.4

psycopg2-binary

djangorestframework>=3.11,<3.12
djangorestframework-simplejwt==4.4.0

requests # ==2.23.0
Pillow # ==7.0.0
sentry-sdk
google-auth==1.14.0
jsonschema
celery==4.3.0